Reorganize app
