+++ bikespace_api/bikespace_api/admin/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+admin_blueprint = Blueprint("admin_routes", __name__, template_folder="templates")
