--- bikespace_api/bikespace_api/db/Dockerfile
@@ -1,5 +1,0 @@
-# base image
-FROM postgres:10.4-alpine
-
-# run create.sql on init
-ADD create.sql /docker-entrypoint-initdb.d
