--- bikespace_api/bikespace_api/db/create.sql
@@ -1,2 +1,0 @@
-CREATE DATABASE bikespace_dev;
-CREATE DATABASE bikespace_test;
