+++ bikespace_api/bikespace_api/submissions/submissions_models.py
@@ -1,11 +1,8 @@
-# bikespace_api/bikespace_api/api/models.py
-
 from datetime import datetime, timezone
 from enum import Enum
 
 import sqlalchemy as sa
 import sqlalchemy.dialects.postgresql as pg
-from flask_security.core import RoleMixin, UserMixin
 from sqlalchemy_continuum import make_versioned
 from sqlalchemy_continuum.plugins import FlaskPlugin
 
@@ -69,36 +66,3 @@ class Submission(db.Model):
         self.parking_time = parking_time
         self.comments = comments
         self.submitted_datetime = datetime.now(timezone.utc)
-
-
-roles_users = db.Table(
-    "roles_users",
-    db.Column("user_id", db.Integer(), db.ForeignKey("user.id")),
-    db.Column("role_id", db.Integer(), db.ForeignKey("role.id")),
-)
-
-
-class Role(db.Model, RoleMixin):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80), unique=True)
-    description = db.Column(db.String(255))
-
-    def __str__(self):
-        return str(self.name or "")
-
-
-class User(db.Model, UserMixin):
-    id = db.Column(db.Integer, primary_key=True)
-    first_name = db.Column(db.String(255))
-    last_name = db.Column(db.String(255))
-    email = db.Column(db.String(255), unique=True)
-    password = db.Column(db.String(255))
-    active = db.Column(db.Boolean())
-    confirmed_at = db.Column(db.DateTime())
-    roles = db.relationship(
-        "Role", secondary=roles_users, backref=db.backref("users", lazy="dynamic")
-    )
-    fs_uniquifier = db.Column(db.String(64), unique=True, nullable=False)
-
-    def __str__(self):
-        return str(self.email or "")
