+++ bikespace_api/bikespace_api/tests/test_home_page.py
@@ -1,7 +1,3 @@
-from bikespace_api import create_app
-import os
-
-
 def test_home_page(test_client):
     """
     GIVEN a Flask application configured for testing
